A bunch of minor tweaks and bug fixes:
1- Yes, I added assert() to FortressLibrary.  I've rewritten it too
many times, and on at least one occasion (as shown by this checkin)
gotten it wrong (such that it always succeeds).  There are three
variants: one takes a Boolean, one takes a Boolean and a diagnostic
string (preferred to the first), and the last takes two values and a
diagnostic string and compares them for equality (preferred to either
of the above because it does toString() on the candidate values iff
the assertion fails).  I cleaned out all the assertion code I could
find in a quick scan of the tests.

Most of said code was my fault.  Please make sure you actually inspect
expression values automatically in test code.

2- Fixed bug (?) in bit shifting by more than the bit width of the
underlying integer type.  We now give 0 for out-of-range left shifts,
and the sign bit for out-of-range right shifts.

3- Made builtin Bools print with a leading lower case letter, just as
they appear in Fortress code.

4- Fixed length-0 bug in argument checking

5- Added a few library comments and did a teensy bit of refactoring.
More is in progress.  This included moving tests/fingerTree to
test_library/PureList where it belongs.