Fixes and code cleanup for applying a single-argument function to a tuple-valued argument (including VOID).  Also cleaned up overloading resolution to do all its checking using the code for tuple subtyping (which is correct) rather than the approximation of the same code that used to live in FType (which was wrong).  Had to fix a couple of workarounds to the bug in the libraries, which had now become ambiguous overloadings which broke messily.
There is still some sort of lingering inconsistency between loosely juxtaposed
application (which handles tuples wrong somehow) and tightly
juxtaposed application (which handles tuple arguments correctly).

There were a bunch of changes to PureList, which are only halfway
complete; I won't vouch for the present state of the code except that
it mostly works as well / badly as it ever did (which isn't too bad,
actually).