Lots of work on Indexed generators, and a good bit of re-factoring to accompany it.  I've now filled out the hierarchy of Ranges (the things you can write with # and :, though complete constructor support is still yet to come).  Next goal is to require all Indexed generators to support indexing with a Range of the Index type, and to clean up various auxiliary libraries to match.
I fixed a bunch of minor interpreter bugs and infelicities along the
way:
* An illegal type instantiation (one that's outside a type
variable's bounds) should fail with a ProgramError, so that it can be
caught if it occurs during unification / overloaded method choice.
Somehow we were never running into this problem before (it seemed to
be a heisenbug depending on the order in which overloadings were
considered).
* When refactoring unifyNonVar to use error(), its type was changed
from boolean to Boolean.  Switched it back as this is just
gratuitous; we just follow calls to error with "return false"
indicating unification failure.
* Catch StackOverflowError around method call, to catch infinite loops
of method calls and report a Fortress stack trace in addition to the
giant Java stack trace.
* Turned a few NYIs into bugs in order to find locations where I was
using still-unimplemented features.