In order to allow files compiled with our "fss" front end to be run with the current interpreter, I added aliases for each primitive type to the top-level environment. Now, in addition to types like:
String, ZZ64, etc.

there are also primitive types like:

FortressBuiltin.String, FortressBuiltin.ZZ64, etc.

This was necessary because the static disambiguator resolves references to those builtin types to things in the FortressBuiltin API. I don't expect that this will affect any existing code (unless you are using names like FortressBuiltin.String for some other purpose, in which case you should fix your code.) :-)

One other thing: In EvalType.java, when an InstantiatedType is evaluated, there was a check to ensure that the type is an instantiation of a FTypeGeneric and, if not, an error is signaled. I disabled this check because it's inconsistent with the static disambiguator. The static disambiguator converts all trait type references to InstantiatedTypes, possibly with zero arguments.