I've done what every functional programmer does when handed a new programming language: I've implemented an interpreter for the lambda calculus.  Actually, there's two extra little frobs: there are top-level free variable bindings (which we use to define a library of combinators); the interpreter attempts to leave top-level variables uninstantiated until they're required for normalization.  Taken together these permit top-level recursion as well (which we see in the definition of fib).
A lot of the complexity here is in the parser, which tries to munch
lambda expressions in a newline-sensitive way (much as Fortress
does).  So if we expect another token, we accept newlines until we've
seen the expected token, but if we're not sure if there will be
another token or not then a newline terminates the expression.
Bindings are simply added to the top-level environment; expressions
are evaluated twice: first to weak head normal form, second to normal
form.  Note that this means that you should not e.g. write "fib" on a
line by itself, since the fib function (being recursive) has no normal
form.

Note that we implement a straight substitution model, so we duplicate
redexes like nobody's business.  Evaluating (fib six) is astonishingly
slow!  That's because we keep having to subtract over and over and
over again.

The test data is found in Lambda.tests.