Additional revisions to permit list comprehensions and the like to be compiled.  This changed the desugarings somewhat.
An intermediate effort to try out the desugaring we *want* proved that
indeed the preferred desugaring will not work without type inference
due to inability to instantiate the element types of intermediate
generators along the way.

Fixed a bug in Reflect that only manifests when this type is used from
multiple unit tests (we were using a stale GenericConstructor to build
fresh instances).