Refactored libraries to move arithmetic operators from top level into the relevant traits and objects.
This turned up a massive oversight in how the "asif" construct was
being handled (and also turned up the fact that in several interpreter
contexts it was being ignored entirely).  Method call has been
re-factored so that all method calls are handled in one place, and
fixed this so that it works (though there are some outstanding
questions noted in the code).