Lots more refactoring to method application in hopes of speeding it up, and to fix some bugs that manifested during the cleanup.
We now push overloaded functional methods all the way through to
obtain a MethodClosure and store that in the overloading table after
overloading resolution.

This doesn't actually appear to help testing time any.  I'm in the
process of an audit to see what else is amiss.  It may be that we're
doing overriding and things rather inefficiently in the libraries.