Spent a while trying to hunt down stuff related to bug #116.  Didn't ultimately figure out how to best solve this problem.  It no longer fails with a null pointer exception, but instead with a type error---which is more like the right thing, but still isn't the "no dotted entity by that name" result we want.
I ended up avoiding the null pointer exception by dropping instances
of `GenericFunctionalMethod` in
`OverloadedFunction.finishInitializingSecondPart`.  It looks a bit
like we're not consistency checking the generic overloading as a
result---but we're never initializing its type information either.
Oddly just dropping the `GenericFunctionalMethod` on the floor doesn't
work either; shouldn't be too shocking, I suppose, but where is the
thing living after we've created the `OverloadedMethod` and we've
dropped it?  The more I look at this, the more puzzling I'm finding
it.

Meanwhile I've done quite a bit of cleanup on method initialization.
Made finishInitializing into a method that works for any Fcn and got
rid of a ton of gratitous instanceof and cast nonsense.  So this
checking at least gets rid of (quite a bit) more code than it adds.