[Libraries] Changed Integral to be a parametric type Integral[\I\].  This hopefully is a start to resolving some problems Guy and I were having integrating unsigned integer types.
Note that this means that the Number hierarchy is no longer strictly
closed (since we can't write a comprises clause that mentions
Integral[\I\]).  As a result you may find that you need to add
"excludes { Number }" to traits involving some of the crazier sorts of
overloading (mostly involving overloadings at numeric and non-numeric
types).  Several tests were changed as a result.

Two bugs were tickled: one was an oversight in exclusion checking (it
turns out checking a subtype condition one-sidedly wasn't good enough,
as our transitive guarantees are quite strong enough for this to find
an obvious exclusion relationship).  The other was a known bug in
static constraint solving, now raising its head in a place where
wasn't previously occurring; this is Ticket #143.