Driver/componentwrapper/desugarer now do a better job of tracking (variable) name uses.
Types still need work.

References from native code depend on a hack in Driver.java;
doing this right requires ensuring that native methods are run in the right environment,
and that the component that specifies the native code also mentions the names referred to
in the native code.

Some clean up of stringName/nameString; more still needs to be done.  Where possible, they
have the same behavior now, so that removing stringName will be less painful and surprising.

Enhanced BASet; the now include the method addArray, and their iterators support remove.