[desugarer] Restructured the FreeNameVisitor to traverse the entire AST corresponding to the component instead of just under ObjectExpression to collect more information I need in order to rewrite mutable varRefs captured by the object expressiond.
I also changed the test case objectCC.fss in the tests directory; it seems
like the useage of operator "+" is causing some problem in the interpreter
when run using JUnit test, but not causing problem when parse / unparse and
run the unparsed version manually.  We are still trying to figure out why
this strange behavior occurs only when run w/ JUnit.