The interpreter is no longer performing imports, and instead relies on information from the static disambiguator.  This remove 50% of the code from Driver, but also required that many implicit references to library symbols be changed to mention the library explicitly.  This may not be correct, if there should have been a local overloading instead (the various nodes with implicit references to symbols usually have an explicit field with the resolved reference, filled in by static checking -- not part of this commit, but if you are tracking down a bug, that's one thing to look for).
In several cases, plain old strings (e.g., "=") were
replaced with references to !WellKnownNames.

Two tests were moved to not_passing_yet; the static end
does not yet deal with renaming imports, and the story for
export of an imported symbol is lacking.  There is preliminary
code to handle this in !ComponentWrapper and !BuildApiEnvironment,
but this may be something that ought to be handled by static
analysis instead (essentially, it is another set of implicit
references, this time at the component level).

There's a minor optimization in the !OverloadedFunction code, too;
the cache was being filled with the non-instantiated version of
generics, which forces repeated unification and instantiation at
later invocations.