[compiler, library, tests] Work on getting RR64 to run (seems OK so far except for the usual warts that also hold for ZZ32; eg literals don't support asString due to dotted method inheritance bugs).  Note that ZZ32 and RR64 arithmetic can't be commingled at all at the moment.
Moved desugaring of loops and generalized if/while to occur before
type checking.  This makes these constructs work the same way as
comprehensions and reductions, which had previously been shifted
earlier in compilation.  Adjusted the desugarings to generate
_RewriteFnApp or MethodApplication nodes rather than TightJuxts.  This
should improve future reliability of desugared code.

Added test of floating-point printing.  Adjusted Compiled7fo to work
around obvious shortcomings in code generation.  Codegen now crashes
on this program, probably due to an outstanding bug that's been
lurking for a long time.  Addressing this is probably the next thing
to be done.