[codegen] Dependency analysis for top-level value bindings (including singletons).
Note that this does not handle inter-module dependencies.  In order to
handle such dependencies correctly it may be necessary to actually
transform this into use of the lazy class initialization pattern.
This would require thoroughly revamping the way all references to
top-level variables work, however, and thus seems undesirable; if it
comes to that we may wish to try to be a bit smarter.