[codegen] Complete rewrite of top-level binding handling.
Previously top level bindings (including singleton objects) were
stored as static fields in the class representing the component.  Now
each binding has its own inner singleton class; its value is stored in
the `ONLY` static field of that class.  For singleton objects the
singleton class is the object's original class, giving them the
intuitive interpretation (and hopefully making singletons play a bit more
nicely with generic instantiation).  For value bindings, class
initialization computes the value of the binding and stores it into
`ONLY`.

This change permits us to handle arbitrary inter-component
dependencies between top-level bindings, and permits top-level binding
computations to create tasks.  This was impossible with the previous
compilation scheme, as the task objects could not call back into the
static methods of the still-being-initialized component class.

Nonetheless, singleton classes are an awfully big hammer, especially
for simple top-level constants.  I have not specialized
`VarCodeGen.StaticBinding` for this use case (it was general enough to
handle it) in case we want to selectively re-introduce local static
fields at the component level for stuff that doesn't require
nontrivial code generation such as static constants and the like.  In
order to do this, though, we'll need to be able to capture such things
in the api (or restrict such an optimization to component-local
bindings).