[rts] Made more work stealing state in the rts static final (initialized on startup); the fact that several constants were not was a significant oversight.  Added a flag to set spawn threshold so that we can play with that in performance studies:
* FORTRESS_SPAWN_THRESHOLD is 5 by default (meaning each thread tries
to keep enough work items around to feed the rest of the threads, plus
5 more).  If the threshold is negative, there is no limit and every
work item is made available to be stolen.  Otherwise varous positive
settings can be experimented with.  This setting is meaningless if
FORTRESS_THREADS=1, as no task is ever spawned in that case.
* FORTRESS_THREAD_STATISTICS=t, as a reminder, will print steal counts,
worker thread pool size, and various other stats as the program exits.
* FORTRESS_THREADS=n continues to set the desired number of worker
threads, and continues to default to the number of cpus on your
system.  Setting FORTRESS_THREADS=1 runs sequentially.