Made all the changes necessary for the Fortress character type to be called "Character" (as in Fortress spec 1.0 beta) rather than "Char".
Doing this revealed an inderlying problem in the NamingCzar foreign
interface mechanism: foreign imports of Java code simply assumed that for
every primitive Java type there is a single corresponding Fortress type;
thus int would be mapped to ZZ32, double to RR64, and char to Char.  The
problem is that Fortress supports full 21-bit Unicode, not just 16-bit
Unicode, and so the Java int type must be used to represent the Fortress
character type.  The good news is, Java now has two versions of (nearly)
all the character functions such as isUpperCase and toLowerCase, one that
works with the Java char type and one that works with the Java int type.
The semi-bad news is, attempting to use the latter means that the
interface on the Fortress side appears to use ZZ32 for argument and result
types.  No problem; all we need is two trivial transfer functions, one to
convert from ZZ32 to Character and one to convert from Character to ZZ32.

Ah, but there's the rub: those two operations are trivial to code in Java
(in nativeHelpers/simpleChar.java), but when they are imported with
"import java", how does the comipler know to use the Character type for
the argument or result type?

The answer is, a magic hack.  The method NamingCzar.fortressTypeForForeignJavaType
now takes three arguments: the Java type descriptor (as before), the method name,
and whether an argument or result type is being processed.  This method now
special-cases two very particular method names:

  charMakeCharacterWithSpecialCompilerHackForCharacterResultType
  charCodePointWithSpecialCompilerHackForCharacterArgumentType

Yes, they're ugly, but I wanted to make sure that no one would stumble
across them by accident.  They are used in just two places: to define the
Java methods in nativeHelpers/simpleChar.java, and in the "import java"
statement in ProjectFortress/LibraryBuiltin/CompilerBuiltin.fss, as follows:

import java com.sun.fortress.nativeHelpers.{
  simpleChar.charCodePointWithSpecialCompilerHackForCharacterArgumentType => jCodePoint,
  simpleChar.charMakeCharacterWithSpecialCompilerHackForCharacterResultType => jMakeCharacter
  }

Then all the Fortress code for trait Character uses these two methods to perform
the necessary type conversions:

    toLowerCase(self): Character = jMakeCharacter(jCharToLowerCase(jCodePoint(self)))

and we hope that optimizations downstream will open-code the transfer methods.